Adds an expo example
