+++ examples/todo-expo/App.tsx
@@ -1,24 +1,28 @@
 import { StatusBar } from 'expo-status-bar'
 import React from 'react'
 import { StyleSheet, Text, View } from 'react-native'
-import { Styles } from './constants'
+import { Styles, SUPABASE_URL, SUPABASE_KEY } from './lib/constants'
 
 import Button from './components/Button'
 import TextField from './components/TextField'
 
 export default function App() {
-  const [email, onChangeEmail] = React.useState('')
-  const [password, onChangePassword] = React.useState('')
+  const [email, setEmail] = React.useState('')
+  const [password, setPassword] = React.useState('')
+
+  console.log(SUPABASE_KEY)
+  
 
   return (
     <View style={styles.container}>
       <View style={styles.verticallySpaced}>
         <Text style={{ fontSize: Styles.fontExtraLarge, fontWeight: 'bold' }}>To Do List</Text>
+        <Text style={{ fontSize: Styles.fontExtraLarge, fontWeight: 'bold' }}>{SUPABASE_KEY}</Text>
       </View>
       <View style={[styles.verticallySpaced, { marginTop: 20 }]}>
         <Text>Email</Text>
         <TextField
-          onChangeText={(text) => onChangeEmail(text)}
+          onChangeText={(text) => setEmail(text)}
           text={email}
           placeholder="Enter your email"
         />
@@ -26,7 +30,7 @@ export default function App() {
       <View style={styles.verticallySpaced}>
         <Text>Password</Text>
         <TextField
-          onChangeText={(text) => onChangePassword(text)}
+          onChangeText={(text) => setPassword(text)}
           text={password}
           type={'password'}
           placeholder="Enter your password"
