+++ examples/todo-expo/babel.config.js
@@ -2,5 +2,6 @@ module.exports = function(api) {
   api.cache(true);
   return {
     presets: ['babel-preset-expo'],
-  };
+    plugins: ['inline-dotenv'],
+  }
 };
