--- examples/todo-expo/constants.ts
@@ -1,8 +1,0 @@
-export const Styles = {
-  fontNormal: 20,
-  fontMedium: 28,
-  fontLarge: 34,
-  fontExtraLarge: 40,
-  colorPrimary: 'black',
-  spacing: 12,
-}
