+++ examples/todo-expo/package.json
@@ -8,6 +8,9 @@
     "eject": "expo eject"
   },
   "dependencies": {
+    "@react-native-community/async-storage": "^1.12.1",
+    "@supabase/supabase-js": "^1.0.2",
+    "babel-plugin-inline-dotenv": "^1.6.0",
     "expo": "~39.0.2",
     "expo-status-bar": "~1.0.2",
     "react": "16.13.1",
